Delete date script
All this file did was reformat the date every time date was ran. I have no idea what the point of this file was (I assume it was for polybar) but it caused a huge breakage in the system, mainly with yay and installing packages.
I don't know why I ever thought that script was a good idea.